+++ FoodTruckVR/Assets/Scenes/CustomHands 1.unity
@@ -38,7 +38,7 @@ RenderSettings:
   m_ReflectionIntensity: 1
   m_CustomReflection: {fileID: 0}
   m_Sun: {fileID: 1882561470}
-  m_IndirectSpecularColor: {r: 0.35239303, g: 0.5784324, b: 0.76843244, a: 1}
+  m_IndirectSpecularColor: {r: 0.3517761, g: 0.57806706, b: 0.76798236, a: 1}
   m_UseRadianceAmbientProbe: 0
 --- !u!157 &3
 LightmapSettings:
@@ -292,7 +292,7 @@ PrefabInstance:
     - target: {fileID: 8134665729841089386, guid: 400e9870199462b4db9fea936a599136,
         type: 3}
       propertyPath: IngredientType
-      value: 5
+      value: 1
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 400e9870199462b4db9fea936a599136, type: 3}
@@ -366,7 +366,7 @@ PrefabInstance:
     - target: {fileID: 8134665729841089386, guid: 400e9870199462b4db9fea936a599136,
         type: 3}
       propertyPath: IngredientType
-      value: 2
+      value: 1
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 400e9870199462b4db9fea936a599136, type: 3}
@@ -440,7 +440,7 @@ PrefabInstance:
     - target: {fileID: 8134665729841089386, guid: 400e9870199462b4db9fea936a599136,
         type: 3}
       propertyPath: IngredientType
-      value: 3
+      value: 1
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 400e9870199462b4db9fea936a599136, type: 3}
