--- FoodTruckVR/Assets/Scripts/Food/BeefGroup.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9d6f0e4313ba3974d9ed40b4c70b8bff
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
