--- FoodTruckVR/Assets/Scripts/Food/CheeseGroup.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7e489ab74a107254583c5ae9fff7693e
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
