+++ FoodTruckVR/Assets/Scripts/Food/IngredientGroup.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a9f2e2a4de743b64a9d0bc7e11b2bb00
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
