--- FoodTruckVR/Assets/Scripts/Food/LettuceGroup.asset
@@ -1,20 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
-  m_Name: LettuceGroup
-  m_EditorClassIdentifier: Assembly-CSharp::IngredientGroup
-  CollectionType: 2
-  IngredientModels:
-  - {fileID: 1165499025449004758, guid: 2f08b4aab4cdf7d41bd220611f618cd4, type: 3}
-  - {fileID: -1275510179500856966, guid: 099d101cd99550446a1c540088d218a2, type: 3}
-  - {fileID: -5645255074800537229, guid: 9ce2c5e21b8ef1d4cad860b94246460f, type: 3}
-  - {fileID: 3900530538902925283, guid: 327701955f351ad4ca2bddcf5422a5d6, type: 3}
