--- FoodTruckVR/Assets/Scripts/Food/LettuceGroup.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: acdf6f1dbaedf944aa1da7e6de8c7e9f
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
