+++ FoodTruckVR/Assets/Scripts/Food/New Ingredient Group.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1a609a77ad28b524b8bf61bd278f0b51
+guid: 50be61df5df418c418273b6dd8c7c081
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
